Move mapping helpers to a file
This increases clarity for finding helpers.
